requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Deep learning and diffusion
diffusers>=0.21.0
transformers>=4.21.0
timm>=0.6.0

# Data processing
PyYAML>=6.0
tqdm>=4.62.0

# Scientific computing
scipy>=1.7.0

# Optional but recommended
matplotlib>=3.5.0
seaborn>=0.11.0
# jupyter>=1.0.0